ai/requirements.txt
tensorflow
fastapi
uvicorn
Pillow
imutils
numpy
opencv-python
keras